Add a type tag in front of every struct
We store all types in the same table. In order to avoid hash collisions between them,
this diff adds a tag in front of every type.
The tag is one byte right now. It means that we're restricted to 256 different types
to store in the DB. I doubt this will be a big issue. An alternative would be
to use bincodes varint encoding that grows with the size of the uint.

I've updated the tests appropriately. I've also deleted one test.