Change the ValueId hash from u64 to [u8;8]
There's a tension when using bincode to encode values and using the u64 type
for hashes. If we uyse the varint encoding, which is good for integers other than hashes,
we waste a byte every hash because it's always a large number. And using the standard encoding
would also be wasteful.
The solution is to use [u8;8] instead for hashes, which doesn't suffer from the varint issues.

Using [u8;8] was suggested in this stackoverflow question:
https://stackoverflow.com/questions/79782435/how-can-i-control-whether-a-u64-is-encoded-as-varint-or-standard-in-bincode

I've cleaned up code a fair bit in the process. I have a single hashing function and
I use the ValueId type more consistently.