Add support for unnamed fields in structs
There's a lot of copy-pasting in this commit, and the code
is arguably longer than it should be. But it's hard to fix it
because numbers are not valid identifiers so we need different
code for when we're indexing in a named struct and an
unnamed one.

I'll try to clean up the code a bit later though.