Complete sharing between tuple structs and named field structs
The two types of struct now share all code that they concevably can,
for generating the TypeVault trait instance.

Next step is to tackle enums. The code in this commit will likely
have to change as a result, so the current state of the code is not
as clean as it could be, because it will change anyway.