Add support for enums
This is quite a bit change and I had to move things around quite a bit.

The main complication with writing this kind of code is that Rust has
different syntax for named and unnamed fields in structs and variants.
I don't think I've managed to nail the right abstraction here to
make it easier to deal with this code. It's a mess right now
and I'd like to clean it up in future commits. But at least it's
working.